task 3: implement fib in bytecode

+++ src/main/java/edu/hw11/task3/SumExample.java
@@ -1,5 +1,5 @@
 package edu.hw11.task3;
 
 public abstract class SumExample {
-    public abstract int calculate(int n);
+    public abstract long calculate(int n);
 }
